refactor: update some files
