+++ src/main/java/com/sistemaf/domain/repository/estoque/stockitem/StockItemRepository.java
@@ -1,18 +1,18 @@
-package com.sistemaf.domain.repository;
+package com.sistemaf.domain.repository.estoque.stockitem;
 
 import com.sistemaf.domain.filter.StockItemFilter;
 import com.sistemaf.domain.model.StockItem;
 import org.springframework.data.domain.Page;
+import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 import java.util.UUID;
 
-public interface StockItemRepository {
+public interface StockItemRepository extends  FindStockItemQuery, JpaRepository<StockItem, UUID> {
     Optional<StockItem> findBySerial(String serial);
 
     StockItem save(StockItem stockItem);
 
     Optional<StockItem> findById(UUID id);
 
-    Page<StockItem> findWithFilter(StockItemFilter stockItemFilter, Object o);
 }
