+++ src/main/java/com/sistemaf/domain/usecases/stock/UpdateStockItemUseCase.java
@@ -5,7 +5,7 @@ import com.sistemaf.domain.exception.BusinessException;
 import com.sistemaf.domain.exception.EntityNotFoundException;
 import com.sistemaf.domain.model.Produto;
 import com.sistemaf.domain.model.StockItem;
-import com.sistemaf.domain.repository.StockItemRepository;
+import com.sistemaf.domain.repository.estoque.stockitem.StockItemRepository;
 import com.sistemaf.domain.repository.estoque.produto.ProdutoRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.beans.BeanUtils;
