+++ src/test/java/com/sistemaf/domain/usecases/stock/AddStockItemUseCaseUnitTest.java
@@ -3,7 +3,7 @@ package com.sistemaf.domain.usecases.stock;
 
 import com.sistemaf.domain.exception.BusinessException;
 import com.sistemaf.domain.model.StockItem;
-import com.sistemaf.domain.repository.StockItemRepository;
+import com.sistemaf.domain.repository.estoque.stockitem.StockItemRepository;
 import com.sistemaf.domain.repository.estoque.produto.ProdutoRepository;
 import org.instancio.Instancio;
 import org.junit.jupiter.api.DisplayName;
